+++ src/mpi_classes/process-information.hh
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstddef>
+
 #include "utils/singleton.hh"
 
 class ProcessInformation : public Singleton<ProcessInformation>
@@ -7,16 +9,17 @@ class ProcessInformation : public Singleton<ProcessInformation>
     friend class Singleton<ProcessInformation>;
 
 public:
-    void set_information(int rank, int n_client, int client_offset, int n_node, int node_offset, int size);
-    bool is_rank_client(int rank) const;
-    bool is_rank_node(int rank) const;
+    void
+    set_information(size_t rank, size_t n_client, size_t client_offset, size_t n_node, size_t node_offset, size_t size);
+    bool is_rank_client(size_t rank) const;
+    bool is_rank_node(size_t rank) const;
 
-    int rank_;
-    int n_client_;
-    int client_offset_;
-    int n_node_;
-    int node_offset_;
-    int size_;
+    size_t rank_;
+    size_t n_client_;
+    size_t client_offset_;
+    size_t n_node_;
+    size_t node_offset_;
+    size_t size_;
 
 private:
     ProcessInformation() = default;
