+++ src/rpc/rpc-query.cc
@@ -2,7 +2,7 @@
 
 #include <utility>
 
-RPCQuery::RPCQuery(int source_rank, RPC::RPC_TYPE type, int term, content_t content)
+RPCQuery::RPCQuery(size_t source_rank, RPC::RPC_TYPE type, int term, content_t content)
     : source_rank_(source_rank)
     , type_(type)
     , term_(term)
